# src/main/resources/api/orders-openapi-spec-v1.yml
openapi: 3.0.3
info:
  title: Order Enricher API
  version: 1.0.0
  description: API for submitting, enriching, and retrieving customer orders.
paths:
  /orders:
    get:
      summary: Retrieve all enriched orders, with optional filtering
      operationId: getOrders
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: string
            example: "CUST-456"
        - name: productId
          in: query
          required: false
          schema:
            type: string
            example: "PROD-A1"
      responses:
        '200':
          description: A list of enriched orders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnrichedOrderResponse'
    post:
      summary: Submit a raw order for enrichment
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created and enriched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrichedOrderResponse'
        '400':
          description: Invalid input provided.
  /orders/{orderId}:
    get:
      summary: Retrieve an enriched order by its ID
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            example: "ORD-2025-0813-001"
      responses:
        '200':
          description: Successful retrieval of the enriched order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrichedOrderResponse'
        '404':
          description: Order not found.

components:
  schemas:
    OrderRequest:
      type: object
      required: [orderId, customerId, productIds, timestamp]
      properties:
        orderId:
          type: string
          example: "ORD-123"
        customerId:
          type: string
          example: "CUST-456"
        productIds:
          type: array
          items:
            type: string
          example: ["PROD-A1", "PROD-B2"]
        timestamp:
          type: string
          format: date-time
          example: "2025-08-13T04:01:00Z"

    EnrichedOrderResponse:
      type: object
      properties:
        orderId:
          type: string
        timestamp:
          type: string
          format: date-time
        customer:
          $ref: '#/components/schemas/Customer'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        totalPrice:
          type: number
          format: double

    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        street:
          type: string
        zip:
          type: string
        country:
          type: string

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
        tags:
          type: array
          items:
            type: string